Server/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
openai==1.12.0
psycopg2==2.9.9
django-filter==23.5
drf-yasg==1.21.7
pytest==8.0.2
pytest-django==4.8.0
black==24.2.0
flake8==7.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9 